Reorganize and refactor files and classes
Major organizational changes
----------------------------

- Made `builder`, `emulator`, and `hardware` folders to clean
  up the root directory and make clear what is hardware-specific.
  Renamed `loihi_api.py` and `loihi_interface.py` and moved them
  to the hardware folder.
- Moved classes related to Loihi inputs from loihi_cx.py to inputs.py
- Renamed CxSimulator -> EmulatorInterface
- Renamed LoihiSimulator -> HardwareInterface
- Renamed CxAxons -> Axon
- Renamed CxGroup -> Compartment
- Renamed CxProbe -> Probe
- Renamed CxSpikeInput -> SpikeInput
- Renamed CxSynapses -> Synapse
- Renamed hardware.api.SpikeInput -> LoihiSpikeInput
- Made a new LoihiInput base class that SpikeInput inherits from.
  In the future, we plan to investigate other ways of providing
  input to a Loihi model.
- Added LoihiBlock
  - This is a new abstraction above the level of Axon, Compartment,
    Probe, and Synapse that serves as the main container for a "block"
    of the Loihi system (i.e., a chunk, region, slice, or segment).
  - Backend developers should think of the Loihi system (Loihi model)
    as a collection of LoihiBlocks and LoihiInputs.
  - A LoihiBlock can have elements across multiple cores/chips.
  - A LoihiBlock contains one Compartment instance, and any number of
    Axon, Probe and Synapse instances.

Major refactoring
-----------------

- Combined CxModel into Model
- Split the EmulatorInterface class into state components for
  Compartment, Synapses, Axons, Probes.
- Moved discretize and validation methods in Loihi classes to
  `discretize.py` and `validate.py` files. Makes the Loihi
  classes (in `block.py`) easier to read, and reduces the number
  of imports needed when discretizing is done in different places.
- Renamed `Synapses.tracing` to `.learning` for clarity.
- Removed unused group attribute from Axon/Synapse.
- Removed unused location attribute from Compartment.
- Synapse validation now calls SynapseFmt validation.

Other changes
-------------

- Removed "cx" from some, but not all, variable/function names.
- Fixed `test_uv_overflow`.
