+++ .gitignore
@@ -26,7 +26,7 @@ htmlcov
 *.analytics/
 *.logs/
 *.plots/
-nengo_loihi/snips/nengo_io.c
+nengo_io.c
 
 # For .ipynb examples
 *.pdf
