+++ nengo_loihi/builder/__init__.py
@@ -1,0 +1,2 @@
+from .builder import Builder, Model
+from . import connection, ensemble, network, node, probe
