+++ nengo_loihi/emulator/__init__.py
@@ -1,0 +1,1 @@
+from .interface import EmulatorInterface
