+++ nengo_loihi/hardware/__init__.py
@@ -1,0 +1,1 @@
+from .interface import HardwareInterface
