+++ nengo_loihi/hardware/tests/test_simulator.py
@@ -1,6 +1,6 @@
 import pytest
 
-from nengo_loihi import loihi_interface
+from nengo_loihi.hardware import interface as hardware_interface
 
 
 class MockNxsdk:
@@ -12,18 +12,18 @@ def test_error_on_old_version(monkeypatch):
     mock = MockNxsdk()
     mock.__version__ = "0.5.5"
 
-    monkeypatch.setattr(loihi_interface, 'nxsdk', mock)
+    monkeypatch.setattr(hardware_interface, 'nxsdk', mock)
     with pytest.raises(ImportError):
-        loihi_interface.LoihiSimulator.check_nxsdk_version()
+        hardware_interface.HardwareInterface.check_nxsdk_version()
 
 
 def test_no_warn_on_current_version(monkeypatch):
     mock = MockNxsdk()
     mock.__version__ = "0.7.0"
 
-    monkeypatch.setattr(loihi_interface, 'nxsdk', mock)
+    monkeypatch.setattr(hardware_interface, 'nxsdk', mock)
     with pytest.warns(None) as record:
-        loihi_interface.LoihiSimulator.check_nxsdk_version()
+        hardware_interface.HardwareInterface.check_nxsdk_version()
     assert len(record) == 0
 
 
@@ -31,6 +31,6 @@ def test_warn_on_future_version(monkeypatch):
     mock = MockNxsdk()
     mock.__version__ = "0.7.6"
 
-    monkeypatch.setattr(loihi_interface, 'nxsdk', mock)
+    monkeypatch.setattr(hardware_interface, 'nxsdk', mock)
     with pytest.warns(UserWarning):
-        loihi_interface.LoihiSimulator.check_nxsdk_version()
+        hardware_interface.HardwareInterface.check_nxsdk_version()
