+++ nengo_loihi/neurons.py
@@ -1,18 +1,18 @@
 import warnings
 
-import numpy as np
-
-import nengo
+from nengo.builder import Builder as NengoBuilder
+from nengo.builder.neurons import build_lif
 from nengo.exceptions import ValidationError
-from nengo.neurons import NeuronType
+from nengo.neurons import LIF, NeuronType, SpikingRectifiedLinear
 from nengo.params import NumberParam
+import numpy as np
 
 
 def loihi_lif_rates(neuron_type, x, gain, bias, dt):
-    # discretize tau_ref as per CxGroup.configure_lif
+    # discretize tau_ref as per Compartment.configure_lif
     tau_ref = dt * np.round(neuron_type.tau_ref / dt)
 
-    # discretize tau_rc as per CxGroup.discretize
+    # discretize tau_rc as per Compartment.discretize
     decay_rc = -np.expm1(-dt/neuron_type.tau_rc)
     decay_rc = np.round(decay_rc * (2**12 - 1)) / (2**12 - 1)
     tau_rc = -dt/np.log1p(-decay_rc)
@@ -41,12 +41,12 @@ def loihi_rates(neuron_type, x, gain, bias, dt):
 
 
 loihi_rate_functions = {
-    nengo.LIF: loihi_lif_rates,
-    nengo.SpikingRectifiedLinear: loihi_spikingrectifiedlinear_rates,
+    LIF: loihi_lif_rates,
+    SpikingRectifiedLinear: loihi_spikingrectifiedlinear_rates,
 }
 
 
-class LoihiLIF(nengo.LIF):
+class LoihiLIF(LIF):
     def rates(self, x, gain, bias, dt=0.001):
         return loihi_lif_rates(self, x, gain, bias, dt)
 
@@ -65,7 +65,7 @@ class LoihiLIF(nengo.LIF):
         refractory_time[spiked_mask] = tau_ref + dt
 
 
-class LoihiSpikingRectifiedLinear(nengo.SpikingRectifiedLinear):
+class LoihiSpikingRectifiedLinear(SpikingRectifiedLinear):
     def rates(self, x, gain, bias, dt=0.001):
         return loihi_spikingrectifiedlinear_rates(self, x, gain, bias, dt)
 
@@ -151,50 +151,22 @@ class NIFRate(NeuronType):
         output[j > 0] = self.amplitude / (self.tau_ref + 1./j[j > 0])
 
 
-# class NIF(NIFRate):
-#     """Spiking version of non-leaky integrate-and-fire (NIF) neuron model.
-
-#     Parameters
-#     ----------
-#     tau_ref : float
-#         Absolute refractory period, in seconds. This is how long the
-#         membrane voltage is held at zero after a spike.
-#     min_voltage : float
-#         Minimum value for the membrane voltage. If ``-np.inf``, the voltage
-#         is never clipped.
-#     amplitude : float
-#         Scaling factor on the neuron output. Corresponds to the relative
-#         amplitude of the output spikes of the neuron.
-#     """
-
-#     probeable = ('spikes', 'voltage', 'refractory_time')
-
-#     min_voltage = NumberParam('min_voltage', high=0)
-
-#     def __init__(self, tau_ref=0.002, min_voltage=0, amplitude=1):
-#         super(NIF, self).__init__(tau_ref=tau_ref, amplitude=amplitude)
-#         self.min_voltage = min_voltage
-
-#     def step_math(self, dt, J, spiked, voltage, refractory_time):
-#         refractory_time -= dt
-#         delta_t = (dt - refractory_time).clip(0, dt)
-#         voltage += J * delta_t
-
-#         # determine which neurons spiked (set them to 1/dt, else 0)
-#         spiked_mask = voltage > 1
-#         spiked[:] = spiked_mask * (self.amplitude / dt)
-
-#         # set v(0) = 1 and solve for t to compute the spike time
-#         t_spike = dt - (voltage[spiked_mask] - 1) / J[spiked_mask]
-
-#         # set spiked voltages to zero, refractory times to tau_ref, and
-#         # rectify negative voltages to a floor of min_voltage
-#         voltage[voltage < self.min_voltage] = self.min_voltage
-#         voltage[spiked_mask] = 0
-#         refractory_time[spiked_mask] = self.tau_ref + t_spike
-
-
 class NIF(NIFRate):
+    """Spiking version of non-leaky integrate-and-fire (NIF) neuron model.
+
+    Parameters
+    ----------
+    tau_ref : float
+        Absolute refractory period, in seconds. This is how long the
+        membrane voltage is held at zero after a spike.
+    min_voltage : float
+        Minimum value for the membrane voltage. If ``-np.inf``, the voltage
+        is never clipped.
+    amplitude : float
+        Scaling factor on the neuron output. Corresponds to the relative
+        amplitude of the output spikes of the neuron.
+    """
+
     probeable = ('spikes', 'voltage', 'refractory_time')
 
     min_voltage = NumberParam('min_voltage', high=0)
@@ -217,11 +189,11 @@ class NIF(NIFRate):
         refractory_time[spiked_mask] = self.tau_ref + dt
 
 
-@nengo.builder.Builder.register(NIFRate)
+@NengoBuilder.register(NIFRate)
 def nengo_build_nif_rate(model, nif_rate, neurons):
-    return nengo.builder.neurons.build_lif(model, nif_rate, neurons)
+    return build_lif(model, nif_rate, neurons)
 
 
-@nengo.builder.Builder.register(NIF)
+@NengoBuilder.register(NIF)
 def nengo_build_nif(model, nif, neurons):
-    return nengo.builder.neurons.build_lif(model, nif, neurons)
+    return build_lif(model, nif, neurons)
