+++ nengo_loihi/tests/test_decode_neurons.py
@@ -2,9 +2,13 @@ import nengo
 import numpy as np
 import pytest
 
-from nengo_loihi.builder import (
-    Model, NoisyDecodeNeurons, OnOffDecodeNeurons,
-    Preset5DecodeNeurons, Preset10DecodeNeurons,
+from nengo_loihi.builder import Model
+from nengo_loihi.decode_neurons import (
+    DecodeNeurons,
+    NoisyDecodeNeurons,
+    OnOffDecodeNeurons,
+    Preset5DecodeNeurons,
+    Preset10DecodeNeurons,
 )
 
 
@@ -107,3 +111,15 @@ def test_node_neurons(decode_neurons, tolerance, Simulator, seed, plt):
     plt.title("error = %0.2e" % error)
 
     assert error < tolerance
+
+
+def test_decode_neuron_str():
+    assert str(DecodeNeurons(dt=0.005)) == "DecodeNeurons(dt=0.005)"
+    assert str(OnOffDecodeNeurons(pairs_per_dim=2, dt=0.002, rate=None)) == (
+        "OnOffDecodeNeurons(pairs_per_dim=2, dt=0.002, rate=250)")
+    assert str(NoisyDecodeNeurons(1, rate=20)) == (
+        "NoisyDecodeNeurons(pairs_per_dim=1, dt=0.001, rate=20, noise_exp=-2)")
+    assert str(Preset5DecodeNeurons()) == (
+        "Preset5DecodeNeurons(dt=0.001, rate=200)")
+    assert str(Preset10DecodeNeurons(dt=0.0001, rate=0.5)) == (
+        "Preset10DecodeNeurons(dt=0.0001, rate=0.5)")
