+++ nengo_loihi/tests/test_splitter.py
@@ -3,14 +3,16 @@ import nengo
 from nengo.exceptions import BuildError
 import numpy as np
 
-from nengo_loihi.builder import OnOffDecodeNeurons
+from nengo_loihi.decode_neurons import OnOffDecodeNeurons
 from nengo_loihi.config import add_params
-from nengo_loihi.neurons import NIF
-from nengo_loihi.splitter import (
+from nengo_loihi.inputs import (
     ChipReceiveNeurons,
     ChipReceiveNode,
     HostReceiveNode,
     HostSendNode,
+)
+from nengo_loihi.neurons import NIF
+from nengo_loihi.splitter import (
     PESModulatoryTarget,
     place_ensembles,
     place_internetwork_connections,
